+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 6.2.2
 ----------------
 - Fix Hacking Log when Parent switching Student in Side.php
 - Do not allow Parents to add New/Existing Address/Contact in Address.inc.php
+- Do not allow Parents to edit Enrollment Records in StudentsUsersInfo.fnc.php
 
 Changes in 6.2.1
 ----------------