+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 6.2.2
 - Fix Hacking Log when Parent switching Student in Side.php
 - Do not allow Parents to add New/Existing Address/Contact in Address.inc.php
 - Do not allow Parents to edit Enrollment Records in StudentsUsersInfo.fnc.php
+- Rollback Give non admin users access to "No Address" in Address.inc.php
 
 Changes in 6.2.1
 ----------------